v0.0.6: M6 Revenue & Dealership Engine — security fixes, RBAC, auth, frontend wiring
Fix 15 verified findings from M6 review:
- SEC001 (Critical): Fix IDOR on payout PDF download — add dealer ownership check
- SEC003 (High): Add requireAdmin middleware to all admin route groups
- SEC002 (High): Remove internal error leakage in dealer approval handler
- SIL001 (High): Check JSON decode error in Suspend handler
- SIL002-004 (High): Add auth tokens to CouponManagement, PayoutApproval, MarginManagement
- PERF001 (High): BulkGenerate now distinguishes duplicate vs real errors
- ARCH001: Add payout status transition validation (pending→approved, processing→paid)
- CODE001: GetHistory now returns all margins including expired (new ListAllMargins)
- CODE002: ListStatements filters by year in SQL instead of in-memory
- CODE004: RejectDealer sets status to "rejected" + added to CHECK constraint
- DB001: Align discount_value column to BIGINT (matches Go int64)
- DB003: Add NOT NULL to pan_number column
- CODE003: Wire Edit button in MarginManagement to open form

Also includes: OAuth Google provider, dashboard pages, new frontend routes,
E2E test specs, logo assets, plugin configs, and component refinements.

+++ backend/internal/database/migrations/027_create_m6_coupon_tables.up.sql
@@ -9,7 +9,7 @@ CREATE TABLE coupons (
     dealer_id UUID REFERENCES dealers(id),
     coupon_type VARCHAR(30) NOT NULL
         CHECK (coupon_type IN ('percentage', 'fixed_amount', 'trial_extension', 'onboarding_bonus', 'feature_addon', 'first_payment_waiver')),
-    discount_value DECIMAL(10,2) NOT NULL,
+    discount_value BIGINT NOT NULL,
     scope_state_id INTEGER REFERENCES states(id),
     scope_plan_id UUID,
     scope_product VARCHAR(50),
